refactor: [US-046] - Move auth views into core/views/auth.py
